remove Cyanogen's setup wizard and install non-market apps by default
Setup wizard asks for wifi password, location sharing and if you want share usage data with CyanogenMod. Only the locale setting would be needed by Replicant. Either it has to be switched to a general setup wizard or this one has to be customized heavily (also the graphics need to be replaced with Replicant ones)

It seems that F-Droid needs this setting enabled again on Android 6

Signed-off-by: Wolfgang Wiedmeyer <wolfgit@wiedmeyer.de>