Remove F-Droid until we can make a FSDG compliant version
It's been more than one year that Replicant is aware that
shipping F-Droid as-is is not compliant with the Free System
Distrbution Guidelines(FSDG)[1].

This is beacause we hoped to resolve the issue upstream in
a timely manner, but that didn't happen.

As we now focus on a longer term solution[2], it doesn't make
any sense to keep F-Droid as-is in Replicant.

That new approach would enable to have an extremely low
maintenance cost for both F-Droid and Replicant if it
is merged upstream.

It would also enable users and developers of other
Android distributions to use it.

References:
-----------
[1]https://www.gnu.org/distros/free-system-distribution-guidelines.html
[2]https://gitlab.com/fdroid/fdroidclient/issues/564#note_162158449

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Joonas Kylmälä <joonas.kylmala@iki.fi>